requirements.txt
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
komodo-python3-dbgp==11.1.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dotenv==0.10.3
Werkzeug==0.15.6
WTForms==2.2.1
